添加font class和unicode

+++ app/service/project.ts
@@ -2,15 +2,15 @@ import { Service } from 'egg'
 import moment = require('moment')
 import uuidv1 = require('uuid/v1')
 
-export default class Icons extends Service {
+export default class PorjectService extends Service {
     public async getList() {
         const SQL = `
-      SELECT *,
-      DATE_FORMAT(create_time, '%Y-%m-%d %T') as create_time,
-      DATE_FORMAT(update_time, '%Y-%m-%d %T') as update_time
-      FROM project
-      order by 'update_time' desc
-    `
+                        SELECT *,
+                        DATE_FORMAT(create_time, '%Y-%m-%d %T') as create_time,
+                        DATE_FORMAT(update_time, '%Y-%m-%d %T') as update_time
+                        FROM project
+                        order by 'update_time' desc
+                    `
 
         // res: Array< RowDataPacket >
         // [
@@ -24,6 +24,7 @@ export default class Icons extends Service {
         const res = await this.app.mysql.query(SQL)
         res.map(item => {
             item.name = item.project_name
+            item.disabled = Boolean(item.disabled)
             delete item.project_name
         })
         return res
@@ -71,18 +72,23 @@ export default class Icons extends Service {
     }
 
     public async update(data: ProjectData) {
+        if (data.id === 'has_removed') {
+            this.ctx.status = 403
+            throw new Error('不可编辑')
+        }
         try {
             const { mysql } = this.app
-            const { id, name } = data
+            const { id, name, fontFace } = data
             const updateTime = moment().format('YYYY-MM-DD HH:mm:ss')
-            const SQL = 'UPDATE project SET project_name=?, update_time=? WHERE id=?'
-            const selectSql = `
-        SELECT *,
-        project_name as name,
-        DATE_FORMAT(create_time, '%Y-%m-%d %T') as create_time,
-        DATE_FORMAT(update_time, '%Y-%m-%d %T') as update_time
-        FROM project WHERE id='${id}'
-      `
+            const SQL = `UPDATE project SET project_name=?, update_time=?, font_face=? WHERE id=?`
+            // const selectSql = `
+            //                     SELECT *,
+            //                     project_name as name,
+            //                     DATE_FORMAT(create_time, '%Y-%m-%d %T') as create_time,
+            //                     DATE_FORMAT(update_time, '%Y-%m-%d %T') as update_time
+            //                     FROM project WHERE id='${id}'
+            //                   `
+            const selectSql = `SELECT * FROM project WHERE id='${data.id}'`
 
             // res:
             // OkPacket {
@@ -96,7 +102,7 @@ export default class Icons extends Service {
             //   changedRows: 1
             // }
 
-            const res = await mysql.query(SQL, [ name, updateTime, id ])
+            const res = await mysql.query(SQL, [ name, updateTime, id, fontFace])
             if (res.affectedRows >= 1) {
                 const [ current ] = await mysql.query(selectSql)
                 delete current.project_name
@@ -109,6 +115,10 @@ export default class Icons extends Service {
     }
 
     public async destroy(id) {
+        if (id === 'has_removed') {
+            this.ctx.status = 403
+            throw new Error('不可删除')
+        }
         try {
             return Promise.all([
                 this.app.mysql.query('DELETE FROM project WHERE id = ?', [ id ]),
